Error message in central video evaluation function has been removed.
Inside a parfor loop, a value (this_target_folder) was assigned inside
of a try statement. This lead to a warning that the variable might not
be assigned in every parfor loop iteration, potentially leading to a
run-time error when value would not be assigned but requested.

Fixed by moving outside of the try statement. Any requests of values
from this_target_folder are still contained within the try statement,
still preventing errors for one video to disrupt the entire analysis of
a batch.